Initial commit of VMS local event flag wakeups.
For each event loop, reserve two VMS local event flags, one for timer wakeups,
and the other for async event wakeups (thread pool completions, poll() wakeups,
AST callback wakeups). The new code is in src/unix/vms-async.c and vms-poll.c.

AST callback completion isn't implemented yet, but will be handled with a
single-linked atomic stack of completed events, which are pushed onto the head
of the (non-threadsafe) async completion queue before handling the entire queue.

Performance of this approach is acceptable on the async1 to async8 benchmarks.
The new poll() thread waits on an additional socketpair() input file descriptor
to received commands from the event loop to restart poll(), pause until the
next call to poll(), or exit the thread. poll() is always called with a timeout
of -1 so that the same poll() can be used across event loop iterations, unless
the list of file descriptor events to watch changes between calls. In that case,
the poll() thread is paused until the next poll() call.

Also fixed descrip.mms to update .exe files when the library has been updated,
and moved the vms "error.c" file into src/unix with the other VMS source files.